requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tqdm>=4.61.0  # Progress bars
tensorboard>=2.5.0  # Training visualization
